feat: Only allow deleting empty dataset (#538)

+++ backend/datasets/delete.py
@@ -1,12 +1,16 @@
 """Delete dataset function."""
 from http import HTTPStatus
 
+import boto3
 from jsonschema import ValidationError, validate  # type: ignore[import]
 from pynamodb.exceptions import DoesNotExist
 
 from ..api_responses import error_response, success_response
 from ..datasets_model import datasets_model_with_meta
+from ..parameter_store import ParameterName, get_param
 from ..types import JsonObject
+
+BOTO3_CLIENT = boto3.client("s3")
 
 
 def delete_dataset(payload: JsonObject) -> JsonObject:
@@ -24,12 +28,22 @@ def delete_dataset(payload: JsonObject) -> JsonObject:
     datasets_model_class = datasets_model_with_meta()
 
     # get dataset to delete
+    dataset_id = req_body["id"]
     try:
-        dataset = datasets_model_class.get(
-            hash_key=f"DATASET#{req_body['id']}", consistent_read=True
-        )
+        dataset = datasets_model_class.get(hash_key=f"DATASET#{dataset_id}", consistent_read=True)
     except DoesNotExist:
-        return error_response(HTTPStatus.NOT_FOUND, f"dataset '{req_body['id']}' does not exist")
+        return error_response(HTTPStatus.NOT_FOUND, f"dataset '{dataset_id}' does not exist")
+
+    # Verify that the dataset is empty
+    storage_bucket_name = get_param(ParameterName.STORAGE_BUCKET_NAME)
+    list_objects_response = BOTO3_CLIENT.list_objects_v2(
+        Bucket=storage_bucket_name, MaxKeys=1, Prefix=f"{dataset_id}/"
+    )
+    if list_objects_response["KeyCount"]:
+        return error_response(
+            HTTPStatus.CONFLICT,
+            f"Can’t delete dataset “{dataset_id}”: dataset versions still exist",
+        )
 
     # delete dataset
     dataset.delete()
